refactor koneksi db
menggabungkan koneksi menjadi satu untuk masing-masing form
