+++ nbproject/private/private.properties
@@ -1,3 +1,3 @@
 compile.on.save=true
 file.reference.mysql-connector-j-8.2.0.jar=C:\\Users\\Ana Bella\\Downloads\\Compressed\\mysql-connector-j-8.2.0\\mysql-connector-j-8.2.0\\mysql-connector-j-8.2.0.jar
-user.properties.file=C:\\Users\\user\\AppData\\Roaming\\NetBeans\\15\\build.properties
+user.properties.file=C:\\Users\\Ana Bella\\AppData\\Roaming\\NetBeans\\19\\build.properties
